// PiggyBank.com-Banking_A/src/app/models/interfaces/inputValidation.model.ts
export interface inputValidationContextItem {
  email: { text1: string; text2: string };
  name: { text1: string; text2: string };
  surname: { text1: string; text2: string };
  cardHolder: { text1: string; text2: string };
  cardNumber: { text1: string; text2: string };
  cardExpdate: { text1: string; text2: string };
  cardCvc: { text1: string; text2: string };
  other: { text1: string; text2: string };
}